+++ utils/activations.py
@@ -1,3 +1,5 @@
+# Activation functions
+
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
